Moving TokenDefinitionState.Value into TokenDefinitionState file
