+++ Source/Stack/4. Layer 3 - CHEMISTRY (Atom builders & reducers)/Reducers/TokenDefinitionsReducer/TokenDefinition.swift
